Add new local random exchange type.
This exchange type will only bind classic queues and will only return
routes for queues that are local to the publishing connection. If more than
one queue is bound it will make a random choice of the locally bound queues.

This exchange type is suitable as a component in systems that run
highly available low-latency RPC workloads.

Co-authored-by: Marcial Rosales <mrosales@pivotal.io>