amqp_channel: Adapt `channel_exit` handling
... after `rabbit_writer` conversion to gen_server.

[Why]
The `rabbit_writer` being a regular gen_server now, it will stop with
the reason `{shutdown, Reason}`. Otherwise this would be considered a
crash by the supervision tree and it would log even more scary errors
and stacktraces.

[How]
If the reason is `{shutdown, _}`, we just unpack the embedded reason and
use it for the channel stop reason. This ensures the behavior of the
channel process doesn't change.